+++ src/Airship/Internal/Parsers.hs
@@ -13,7 +13,7 @@ import Data.Attoparsec.ByteString.Char8 (Parser, parseOnly, sepBy', char,
                                          takeWhile1, inClass, endOfInput)
 import Data.ByteString (ByteString)
 
-import Airship.Internal.Types (ETag(..))
+import Airship.Types (ETag(..))
 
 comma :: Parser Char
 comma = char ','
